Move test files to ./test/ directory, add setup.py and modify imports
