+++ test/setup.py
@@ -1,0 +1,6 @@
+# Add the parent directory to the Python path
+import sys
+sys.path.append('..')
+
+# Import the app module from app.py
+from app import app
